+++ frontend/src/lib/components/Charts/BarChart_horizontal.svelte
