+++ frontend/src/lib/components/Charts/BarChart_vertical.svelte
