+++ frontend/src/lib/components/Charts/LineChart.svelte
